Modularize and refactor code

+++ app/src/main/java/com/koustuvsinha/benchmarker/databases/DbSQLiteHelper.java
@@ -26,8 +26,10 @@ public class DbSQLiteHelper extends SQLiteOpenHelper implements DbTestInterface 
     public static final String COLUMN_ADDRESS = "address";
     public static final String COLUMN_AGE = "age";
 
-    private static final String DATABASE_NAME = "testing.db";
+    private static final String DATABASE_NAME = "testingDb.db";
     private static final int DATABASE_VERSION = 1;
+
+    private SQLiteDatabase db;
 
     //Database creation SQL statement
     private static final String DATABASE_CREATE = "create table " + TABLE_TEST
@@ -42,12 +44,17 @@ public class DbSQLiteHelper extends SQLiteOpenHelper implements DbTestInterface 
 
     public DbSQLiteHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
+        db = this.getWritableDatabase();
     }
 
     @Override
     public void onCreate(SQLiteDatabase sqLiteDatabase) {
         sqLiteDatabase.execSQL(DATABASE_CREATE);
         Log.i(Constants.APP_NAME,"SQLite database created..");
+    }
+
+    public void onDestroy(SQLiteDatabase sqLiteDatabase) {
+
     }
 
     @Override
@@ -61,7 +68,7 @@ public class DbSQLiteHelper extends SQLiteOpenHelper implements DbTestInterface 
 
     public void insertData(List<DbTestRecordModel> modelList) {
         Log.i(Constants.APP_NAME,"Beginning SQLite data insertion..");
-        SQLiteDatabase db = this.getWritableDatabase();
+
         ContentValues values;
         db.beginTransaction();
 
@@ -76,8 +83,6 @@ public class DbSQLiteHelper extends SQLiteOpenHelper implements DbTestInterface 
         }
         db.setTransactionSuccessful();
         db.endTransaction();
-
-        db.close();
         Log.i(Constants.APP_NAME,"SQLite data insertion end");
     }
 
@@ -111,8 +116,14 @@ public class DbSQLiteHelper extends SQLiteOpenHelper implements DbTestInterface 
     }
 
     public void deleteAllData() {
-        SQLiteDatabase db = this.getWritableDatabase();
-        db.execSQL("delete from " + TABLE_TEST);
-        db.close();
+        if(db!=null) {
+            db.execSQL("delete from " + TABLE_TEST);
+        }
+    }
+
+    public void closeDb() {
+        if(db!=null) {
+            db.close();
+        }
     }
 }
